Implement comprehensive SEO improvements and server-side rendering
