+++ app/robots.ts
@@ -1,0 +1,12 @@
+import { MetadataRoute } from 'next';
+
+export default function robots(): MetadataRoute.Robots {
+  return {
+    rules: {
+      userAgent: '*',
+      allow: '/',
+      disallow: ['/dashboard/', '/api/'],
+    },
+    sitemap: 'https://pramit.gg/sitemap.xml',
+  };
+}
