+++ app/sitemap.ts
@@ -1,0 +1,38 @@
+import { getAllPosts } from './lib/server-utils';
+import { MetadataRoute } from 'next';
+
+export default async function sitemap(): Promise<MetadataRoute.Sitemap> {
+  const posts = await getAllPosts();
+  
+  // Static pages
+  const staticPages = [
+    {
+      url: 'https://pramit.gg',
+      lastModified: new Date(),
+      changeFrequency: 'daily' as const,
+      priority: 1,
+    },
+    {
+      url: 'https://pramit.gg/about',
+      lastModified: new Date(),
+      changeFrequency: 'monthly' as const,
+      priority: 0.8,
+    },
+    {
+      url: 'https://pramit.gg/music',
+      lastModified: new Date(),
+      changeFrequency: 'weekly' as const,
+      priority: 0.7,
+    },
+  ];
+
+  // Dynamic post pages
+  const postPages = posts.map((post) => ({
+    url: `https://pramit.gg/post/${post.slug}`,
+    lastModified: new Date(post.updated_at),
+    changeFrequency: 'weekly' as const,
+    priority: 0.9,
+  }));
+
+  return [...staticPages, ...postPages];
+}
